Removes Diameter Stack performance bottleneck at registerRequestRouteInfo
Update Issue 2154
Status: Fixed
By replacing iteration of list containing ordered request info and removing it both from a HashMap and from a CopyOnWriteArrayList by removing a first-half subList from the HashMap keySet and creating a new CopyOnWriteArrayList with the second-half subList, the time wasted for this method has been greatly improved, mainly on the latter.
This method should not limit performance anymore.



git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/diameter@15463 bf0df8d0-2c1f-0410-b170-bd30377b63dc