Fixing incorrect Origin-State-Id value increment.
Update Issue 2291
Status: Fixed
Origin-State-Id was being obtained through System.currentTimeMillis() converted to int, which overflowed and resulted in a negative number to which a Math.abs() was applied. This obviously (due to the overflow) caused the resulting value to effectively decrease instead of increasing as demanded by RFC. The simple fix is to use long instead as it is the expected value anyway since it is of type Unsigned32.

Can you please confirm the fix works OK for you?

I've tried on my end and it looks fine now:
1st CER, Origin-State-Id value: 3125147661
2nd CER, Origin-State-Id value: 3125279425
3rd CER, Origin-State-Id value: 3125329243

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/diameter@16669 bf0df8d0-2c1f-0410-b170-bd30377b63dc