Add possibility of storing routing info at answer
If setting REQUEST_TABLE_SIZE to 0 in configuration, then the answer
routing info extracted from request is stored directly in the answer
object.

This avoids the complex garbage collecting of such table, but developer
must be aware to use the answers created from requests (which happens
naturally), and not creating a raw message to answer received requests.